move pde problems to their own files
